# api/openapi.yaml
openapi: 3.0.3
info:
  title: Propulse Backend API
  version: 0.1.0
paths:
  /api/v1/access:
    get:
      summary: Validate the configured personal access token
      security:
        - AccessBearerAuth: []
      responses:
        '200':
          description: Access token accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessStatusResponse'
        '401':
          $ref: '#/components/responses/AccessRequired'
  /api/v1/capacity/assumptions:
    get:
      summary: Get default calculation assumptions
      description: Public. Returns the complete currently effective rule set used to pre-fill reproducible calculations.
      parameters:
        - name: city
          in: query
          schema:
            type: string
            default: 天津
        - name: homePurchaseOrder
          in: query
          schema:
            type: string
            enum: [first, second]
            default: first
        - name: loanTermMonths
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 360
            default: 360
      responses:
        '200':
          description: Default assumptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapacityAssumptionsResponse'
  /api/v1/capacity/calculations:
    post:
      summary: Create a housing capacity calculation
      security:
        - AccessBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HousingCapacityInput'
      responses:
        '201':
          description: Calculation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCalculationResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Selected asset or target listing not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '409':
          description: Selected target listing is no longer active
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '401':
          $ref: '#/components/responses/AccessRequired'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/capacity/calculations/{id}:
    get:
      summary: Get a saved housing capacity calculation
      security:
        - AccessBearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Calculation found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResponse'
        '404':
          description: Calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/AccessRequired'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/assets:
    post:
      summary: Create a property asset for the configured user
      security:
        - AccessBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateAssetRequest'}
      responses:
        '201':
          description: Asset created with an authoritative source snapshot
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AssetResponse'}
        '400':
          description: Invalid asset facts or request shape
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '401': {$ref: '#/components/responses/AccessRequired'}
        '404':
          description: Selected neighborhood or listing not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '409':
          description: Selected listing is no longer active
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Asset could not be stored
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    get:
      summary: List active property assets owned by the configured user
      security:
        - AccessBearerAuth: []
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        '200':
          description: Active assets ordered by most recently updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AssetsPageResponse'}
        '400':
          description: Invalid pagination
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '401': {$ref: '#/components/responses/AccessRequired'}
        '500':
          description: Asset list could not be read
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/v1/assets/{id}:
    get:
      summary: Get one active property asset owned by the configured user
      security:
        - AccessBearerAuth: []
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Asset found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AssetResponse'}
        '400':
          description: Invalid asset ID
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '401': {$ref: '#/components/responses/AccessRequired'}
        '404':
          description: Asset not found for this user
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Asset could not be read
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    patch:
      summary: Update confirmed facts for one owned property asset
      security:
        - AccessBearerAuth: []
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateAssetRequest'}
      responses:
        '200':
          description: Asset updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AssetResponse'}
        '400':
          description: Invalid asset facts or request shape
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '401': {$ref: '#/components/responses/AccessRequired'}
        '404':
          description: Asset or selected source not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '409':
          description: Selected replacement listing is no longer active
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Asset could not be updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    delete:
      summary: Soft-delete one owned property asset
      security:
        - AccessBearerAuth: []
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '204': {description: Asset soft-deleted}
        '400':
          description: Invalid asset ID
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '401': {$ref: '#/components/responses/AccessRequired'}
        '404':
          description: Asset not found for this user
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Asset could not be deleted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/v1/neighborhoods:
    get:
      summary: Search neighborhoods
      description: Public, anonymous. Returns only neighborhoods with a trusted city and at least one catalog layout. Filter options are not truncated by pagination.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: area
          in: query
          required: false
          schema:
            type: string
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: targetLayout
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Matching neighborhoods
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeighborhoodSearchResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a neighborhood
      security:
        - AccessBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNeighborhoodRequest'
      responses:
        '201':
          description: Neighborhood created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeighborhoodResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/AccessRequired'
  /api/v1/neighborhoods/{id}:
    get:
      summary: Get a neighborhood
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Neighborhood found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeighborhoodResponse'
        '404':
          description: Neighborhood not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/neighborhoods/{id}/community-market:
    get:
      summary: Get the latest source-native community aggregate market snapshot
      description: Public. Aggregate observations and optional source-native profile fields remain separate from listing and transaction detail metrics.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Latest community aggregate snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityMarketSnapshot'
        '404':
          description: Community aggregate snapshot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Snapshot lookup failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/neighborhoods/{id}/community-market/latest:
    get:
      summary: Get the latest complete or aggregate-only community market snapshot
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Latest community market snapshot
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommunityMarketSnapshot'}
        '404':
          description: Snapshot not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Snapshot lookup failed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/v1/neighborhoods/{id}/market-listings:
    get:
      summary: List the latest complete Fangjian listing inventory
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: layout, in: query, schema: {type: string}}
        - {name: floor, in: query, schema: {type: string, enum: [高楼层, 中楼层, 低楼层]}}
        - {name: minPriceWan, in: query, schema: {type: number, minimum: 0}}
        - {name: maxPriceWan, in: query, schema: {type: number, minimum: 0}}
        - {name: sortBy, in: query, schema: {type: string, enum: [date, price, unitPrice, area, adjustments], default: date}}
        - {name: sortOrder, in: query, schema: {type: string, enum: [asc, desc], default: desc}}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        '200':
          description: Paginated current listings
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MarketListingsPage'}
        '400':
          description: Invalid filters
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Complete listing snapshot not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Listing lookup failed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/v1/neighborhoods/{id}/market-listings/{roomId}:
    get:
      summary: Get one listing from the latest complete active inventory
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: roomId, in: path, required: true, schema: {type: string, maxLength: 128}}
      responses:
        '200':
          description: Current authoritative listing and collection metadata
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MarketListingDetail'}
        '400':
          description: Invalid room or neighborhood identifier
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Listing does not exist or is no longer in the latest active inventory
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Listing lookup failed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/v1/neighborhoods/{id}/market-transactions:
    get:
      summary: List the latest complete Fangjian transaction history
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: layout, in: query, schema: {type: string}}
        - {name: floor, in: query, schema: {type: string, enum: [高楼层, 中楼层, 低楼层]}}
        - {name: minPriceWan, in: query, schema: {type: number, minimum: 0}}
        - {name: maxPriceWan, in: query, schema: {type: number, minimum: 0}}
        - {name: sortBy, in: query, schema: {type: string, enum: [date, price, unitPrice, area, adjustments], default: date}}
        - {name: sortOrder, in: query, schema: {type: string, enum: [asc, desc], default: desc}}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        '200':
          description: Paginated transactions
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MarketTransactionsPage'}
        '400':
          description: Invalid filters
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Complete transaction snapshot not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Transaction lookup failed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/v1/neighborhoods/{id}/market-listings/{roomId}/adjustments:
    get:
      summary: List price adjustments for one room in the latest complete snapshot
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: roomId, in: path, required: true, schema: {type: string, maxLength: 128}}
      responses:
        '200':
          description: Price adjustment timeline
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ListingAdjustmentsResponse'}
        '400':
          description: Invalid room or neighborhood identifier
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Complete listing snapshot not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Adjustment lookup failed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/v1/community-market/comparison:
    get:
      summary: Compare the latest market snapshot for two neighborhoods
      parameters:
        - {name: neighborhoodId, in: query, required: true, schema: {type: string, format: uuid}}
        - {name: peerNeighborhoodId, in: query, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Two-neighborhood market comparison
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommunityMarketComparison'}
        '400':
          description: Invalid or identical neighborhood identifiers
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: A comparison snapshot was not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Comparison failed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /api/v1/watchlist/items:
    post:
      summary: Add a neighborhood to the user's watchlist
      security:
        - AccessBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddWatchlistItemRequest'
      responses:
        '201':
          description: Watchlist item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddWatchlistItemResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/AccessRequired'
        '404':
          description: Neighborhood not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The neighborhood is already on the current user's watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: watchlist_item_exists
                  message: the neighborhood is already on the watchlist
  /api/v1/watchlist:
    get:
      summary: List the user's watchlist
      security:
        - AccessBearerAuth: []
      responses:
        '200':
          description: Watchlist items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistResponse'
        '401':
          $ref: '#/components/responses/AccessRequired'
...
  /admin/api/imports:
    get:
      summary: List traceable listing and transaction collection runs
      security:
        - AccessBearerAuth: []
      parameters:
        - {name: dataSourceId, in: query, schema: {type: string, format: uuid}}
        - {name: neighborhoodId, in: query, schema: {type: string, format: uuid}}
        - {name: status, in: query, schema: {type: string, enum: [completed]}}
        - {name: metricStatus, in: query, schema: {type: string, enum: [pending, completed, failed]}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
        - name: page
          in: query
          schema: {type: integer, minimum: 1, default: 1}
        - name: pageSize
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Paginated collection runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionRunsPage'
        '400':
          description: Invalid filter or pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/AccessRequired'
        '500':
          description: Import list failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/api/imports/csv/template:
    get:
      summary: Download the canonical UTF-8 CSV import template
      security:
        - AccessBearerAuth: []
      responses:
        '200':
          description: CSV header template
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            text/csv:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/AccessRequired'
...
components:
  securitySchemes:
    AccessBearerAuth:
      type: http
      scheme: bearer
  responses:
    AccessRequired:
      description: Valid bearer access token is required
      headers:
        WWW-Authenticate:
          description: Authentication scheme required to access this operation
          schema:
            type: string
            example: Bearer
          example: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthorized
              message: valid bearer access token is required
  schemas:
...
    CreateCalculationResponse:
      allOf:
        - $ref: '#/components/schemas/CalculationResponse'
    CalculationResponse:
      type: object
      required: [id, input, result, createdAt]
      properties:
        id:
          type: string
        input:
          $ref: '#/components/schemas/HousingCapacityInput'
        result:
          $ref: '#/components/schemas/HousingCapacityResult'
        selectionContext:
          $ref: '#/components/schemas/PropertySelectionContext'
        createdAt:
          type: string
          format: date-time
...
    AccessStatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [unlocked]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string